ref(actions): Extract resolve actions component from group details view
